Refactor evaluator episode supervisor boundary
